+++ app/src/main/java/com/sopt/now/compose/data/api/service/AuthService.kt
@@ -4,10 +4,10 @@ import com.sopt.now.compose.data.api.ApiKeyStorage.INFO
 import com.sopt.now.compose.data.api.ApiKeyStorage.JOIN
 import com.sopt.now.compose.data.api.ApiKeyStorage.LOGIN
 import com.sopt.now.compose.data.api.ApiKeyStorage.MEMBER
-import com.sopt.now.compose.data.request.RequestLoginDto
-import com.sopt.now.compose.data.request.RequestSignUpDto
-import com.sopt.now.compose.data.response.BaseResponse
-import com.sopt.now.compose.data.response.ResponseMemberInfo
+import com.sopt.now.compose.data.dto.request.RequestLoginDto
+import com.sopt.now.compose.data.dto.request.RequestSignUpDto
+import com.sopt.now.compose.data.dto.response.BaseResponse
+import com.sopt.now.compose.data.dto.response.ResponseMemberInfo
 import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.GET
