+++ app/src/main/java/com/sopt/now/compose/data/api/service/ReqresService.kt
@@ -1,6 +1,6 @@
 package com.sopt.now.compose.data.api.service
 
-import com.sopt.now.compose.data.response.ResponseReqresDto
+import com.sopt.now.compose.data.dto.response.ResponseReqresDto
 import retrofit2.Response
 import retrofit2.http.GET
 import retrofit2.http.Query
