+++ app/src/main/java/com/sopt/now/compose/data/dto/response/BaseResponse.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.compose.data.response
+package com.sopt.now.compose.data.dto.response
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
