+++ app/src/main/java/com/sopt/now/compose/feature/search/SearchScreen.kt
@@ -16,8 +16,8 @@ import androidx.compose.ui.platform.LocalLifecycleOwner
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.sopt.now.compose.component.row.ReqresDummyRow
-import com.sopt.now.compose.core.view.UiState
+import com.sopt.now.compose.ui.component.row.ReqresDummyRow
+import com.sopt.now.compose.ui.core.view.UiState
 
 @Composable
 fun SearchScreen() {
