+++ app/src/main/java/com/sopt/now/compose/feature/search/SearchViewModel.kt
@@ -2,7 +2,7 @@ package com.sopt.now.compose.feature.search
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.sopt.now.compose.core.view.UiState
+import com.sopt.now.compose.ui.core.view.UiState
 import com.sopt.now.compose.data.api.ApiFactory
 import com.sopt.now.compose.feature.model.ReqresEntity
 import kotlinx.coroutines.flow.MutableStateFlow
